update: throwing entity not found exception when updating and deleting courses by id

+++ src/main/java/com/youssef/jpatsk/service/CourseServiceImpl.java
@@ -4,11 +4,13 @@ import com.youssef.jpatsk.dao.CourseRepository;
 import com.youssef.jpatsk.dto.CourseDto;
 import com.youssef.jpatsk.dto.DtoMapper;
 import com.youssef.jpatsk.entity.Course;
+import com.youssef.jpatsk.exception.EntityNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CourseServiceImpl implements CourseService {
@@ -40,11 +42,21 @@ public class CourseServiceImpl implements CourseService {
         Course courseEntity = mapper.dtoToCourse(course);
         courseEntity.setId(id);
 
+        Optional<Course> result = courseRepository.findById(id);
+
+        if (result.isEmpty())
+            throw new EntityNotFoundException("Course id - " + id + " not found");
+
         return mapper.courseToDto(courseRepository.save(courseEntity));
     }
 
     @Override
     public void delete(Long id) {
+        Optional<Course> result = courseRepository.findById(id);
+
+        if (result.isEmpty())
+            throw new EntityNotFoundException("Course id - " + id + " not found");
+
         courseRepository.deleteById(id);
     }
 }
